requirements.txt
requests
telegram
python-telegram-bot
google-cloud-vision
watchdog
